+++ src/haddock/modules/scoring/emscoring/__init__.py
@@ -11,7 +11,7 @@ from haddock.cns.util import (
 from haddock.libs.libparallel import Scheduler
 from haddock.libs.libsubprocess import CNSJob
 from haddock.modules import BaseHaddockModule
-from haddock.ontology import Format, ModuleIO, PDBFile
+from haddock.libs.libontology import Format, ModuleIO, PDBFile
 
 
 logger = logging.getLogger(__name__)
